Add documentation to new components
