--- src/containers/Interfaces/NameGeneratorList/helpers.js
@@ -1,20 +1,0 @@
-/* eslint-disable import/prefer-default-export */
-import { getEntityAttributes } from '../../../ducks/modules/network';
-import getParentKeyByNameValue from '../../../utils/getParentKeyByNameValue';
-
-export const detailsWithVariableUUIDs = (props) => (node) => {
-  const {
-    nodeTypeDefinition,
-    visibleSupplementaryFields,
-  } = props;
-
-  const nodeTypeVariables = nodeTypeDefinition.variables;
-  const attrs = getEntityAttributes(node);
-  const fields = visibleSupplementaryFields;
-  const withUUIDReplacement = fields.map((field) => ({
-    ...field,
-    variable: getParentKeyByNameValue(nodeTypeVariables, field.variable),
-  }));
-
-  return withUUIDReplacement.map((field) => ({ [field.label]: attrs[field.variable] }));
-};
