+++ src/containers/SearchableList/useSortableProperties.js
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { useSelector } from 'react-redux';
-import { get } from 'lodash';
+import { get, compact } from 'lodash';
 import { getProtocolCodebook } from '../../selectors/protocol';
 
 export const getVariableMap = (state) => {
@@ -17,17 +17,21 @@ export const getVariableMap = (state) => {
     );
 };
 
-const withVariableId = (variableMap) => (item) => {
+const withVariableId = (variableMap, path) => (item) => {
   const ref = variableMap.find(([, name]) => name === item.variable);
   const variable = ref ? ref[0] : item.variable;
+
   return {
     ...item,
-    variable,
+    variable: compact([path, variable]),
   };
 };
 
-// Sortable properties with ids
-const useSortableProperties = (sortOptions) => {
+/**
+ * Convert protocol config options into a format
+ * usable by useSort. Essentially specific to SearchableList.
+ */
+const useSortableProperties = (sortOptions, path) => {
   const variableMap = useSelector(getVariableMap);
   const sortableProperties = get(sortOptions, 'sortableProperties');
   const initialSortOrder = get(sortOptions, ['sortOrder', 0]);
@@ -42,7 +46,7 @@ const useSortableProperties = (sortOptions) => {
       const property = ref ? ref[0] : initialSortOrder.property;
       return {
         ...initialSortOrder,
-        property,
+        property: compact([path, property]),
       };
     },
     [initialSortOrder],
@@ -51,7 +55,7 @@ const useSortableProperties = (sortOptions) => {
   const enhancedSortableProperties = useMemo(
     () => {
       if (!sortableProperties) { return []; }
-      return sortableProperties.map(withVariableId(variableMap));
+      return sortableProperties.map(withVariableId(variableMap, path));
     },
     [sortableProperties],
   );
