+++ src/styles/containers/_searchable-list.scss
@@ -51,17 +51,6 @@
     }
   }
 
-  // &__status {
-  //   position: absolute;
-  //   top: 0;
-  //   left: 0;
-  //   width: 100%;
-  //   height: 100%;
-  //   display: flex;
-  //   justify-content: center;
-  //   align-items: center;
-  // }
-
   &__sort {
     padding: 1rem;
     position: absolute;
